requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
gdown==4.5.1
pandas==1.5.0
snowflake-connector-python==2.6.2
pytest
pytest-mock
pyarrow<5.1.0,>=5.0.0